Do *not* cast to a long int!

+++ xpdev/genwrap.c
@@ -2,7 +2,7 @@
 
 /* General cross-platform development wrappers */
 
-/* $Id: genwrap.c,v 1.62 2005/11/29 20:04:27 deuce Exp $ */
+/* $Id: genwrap.c,v 1.63 2005/11/29 20:15:23 deuce Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -515,7 +515,7 @@ long double	DLLCALL	xp_timer(void)
 		ret /= ((long double)freq.HighPart*4294967296)+((long double)freq.LowPart);
 #else
 		/* In MSVC, a long double does NOT have 19 decimals of precision */
-		ret=((long int)(tick.QuadPart/freq.QuadPart))
+		ret=(tick.QuadPart/freq.QuadPart)
 				+(((long double)(tick.QuadPart%freq.QuadPart))/freq.QuadPart);
 #endif
 	}
